Revert "Separate template func"
