+++ src/tools/create_mulmo_script_from_url.ts
@@ -9,8 +9,7 @@ import * as agents from "@graphai/vanilla";
 import { fileWriteAgent } from "@graphai/vanilla_node_agents";
 import { browserlessAgent } from "@graphai/browserless_agent";
 import validateSchemaAgent from "../agents/validate_schema_agent.js";
-import { mkdir, writingMessage } from "../utils/file.js";
-import { readTemplatePrompt } from "../utils/templates.js";
+import { readTemplatePrompt, mkdir, writingMessage } from "../utils/file.js";
 import { browserlessCacheGenerator } from "../utils/filters.js";
 import { mulmoScriptSchema, urlsSchema } from "../types/schema.js";
 import { ScriptingParams } from "../types/index.js";
